third_party/WebKit/LayoutTests/fast/forms/associatedFormControls-leak-nodes.html

<!DOCTYPE html>
<html>
<body>
<iframe id="frame"></iframe>

<script src="../../resources/js-test.js"></script>
<script>
if (!window.internals) {
    debug("This test only runs on \"content_shell --run-layout-test\", as it requires existence of window.internals.");
} else {
    testRunner.waitUntilDone();
    window.jsTestIsAsync = true;
    var documentsBefore;
    var documentsAfter;
    // FIXME(keishi): Calling asyncGC twice to fix flakiness.
    asyncGC(function() {
        asyncGC(function() {
            documentsBefore = window.internals.numberOfLiveDocuments();

            var frame = document.getElementById('frame');
            frame.contentDocument.body.innerHTML = '<form></form>';
            document.body.removeChild(frame);
            frame = null;

            asyncGC(function() {
                documentsAfter = window.internals.numberOfLiveDocuments();

                // -1 is from removing frame itself.
                shouldBe('documentsBefore - 1', 'documentsAfter');
                finishJSTest();
            });
        });
    });
}
</script>
</body>
</html>
